+++ packages/clawql-k8s-operator/src/watches/index.ts
@@ -1,5 +1,6 @@
 export * from "./burst-watch-stub.js";
 export * from "./burst-watch-loop.js";
+export * from "./burst-watch-sources.js";
 export * from "./pod-informer.js";
 export * from "./istio-denial-adapter.js";
 export * from "./istio-access-log-tail.js";